check_files did not support non-path segmentation column
The check_files option assumed that the segmentation column would be a path, causing an error when the segmentation is stored as ordered pairs of coordinates.

I fixed it by checking the type of the first item in the segmentation column. If it is a string, then it checks to make sure it is a valid file path.

We don't currently support path object types, but if we do in the future, we will need to also check that the values are not paths.